+++ pytest_bdd/feature.py
@@ -37,7 +37,7 @@ class FeatureError(Exception):
 
     """Feature parse error."""
 
-    message = u'{0}.\nLine number: {1}.\nLine: {2}.'
+    message = u'{0}.\nLine number: {1}.\nLine: {2}.\nFile: {3}'
 
     def __str__(self):
         return unicode(self).encode('utf-8')
@@ -198,16 +198,16 @@ class Feature(object):
                 if mode == types.GIVEN and prev_mode not in (
                         types.GIVEN, types.SCENARIO, types.SCENARIO_OUTLINE, types.BACKGROUND):
                     raise FeatureError('Given steps must be the first in withing the Scenario',
-                                       line_number, clean_line)
+                                       line_number, clean_line, filename)
 
                 if mode == types.WHEN and prev_mode not in (
                         types.SCENARIO, types.SCENARIO_OUTLINE, types.GIVEN, types.WHEN):
                     raise FeatureError('When steps must be the first or follow Given steps',
-                                       line_number, clean_line)
+                                       line_number, clean_line, filename)
 
                 if not background and mode == types.THEN and prev_mode not in (types.GIVEN, types.WHEN, types.THEN):
                     raise FeatureError('Then steps must follow Given or When steps',
-                                       line_number, clean_line)
+                                       line_number, clean_line, filename)
 
                 if mode == types.FEATURE:
                     if prev_mode != types.FEATURE:
@@ -290,6 +290,7 @@ class Scenario(object):
         self.example_converters = example_converters
         self.tags = tags or set()
         self.failed = False
+        self.test_function = None
 
     def add_step(self, step_name, step_type, indent, line_number, keyword):
         """Add step to the scenario.