+++ pytest_bdd/__init__.py
@@ -1,6 +1,6 @@
 """pytest-bdd public API."""
 
-__version__ = '2.4.0'
+__version__ = '2.4.1'
 
 try:
     from pytest_bdd.steps import given, when, then