+++ CHANGES.rst
@@ -69,7 +69,7 @@ Changelog
 
 
 0.6.11
------
+------
 
 - Fixed step arguments conflict with the fixtures having the same name (olegpidsadnyi)